+++ src/util/hangoutTypes.d.ts
@@ -20,6 +20,7 @@ export interface HangoutEvent extends RowDataPacket {
 
 export interface HangoutMember extends RowDataPacket {
   hangout_member_id: number,
+  username: string,
   user_type: 'account' | 'guest',
   display_name: string,
   is_leader: boolean,